Rename selector to match function name.
